rename_rev: initial commit
This is a script that I've had for a while.

I use it to review patches to drivers/staging.  What it tries to do is strip away the
repetitive stuff and the whitespace changes to see anything important about the
patch.

So if someone sends a patch to rename CamelCase variables then I do:
`cat email.txt | rename_rev.pl -a` and it shows a summary of which variables
were renamed.  It doesn't work perfectly or even very well, but sometimes it
works well enough to be useful.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>